Report Repository/Service

+++ src/main/java/edu/uth/nurseborn/dtos/ReportDTO.java
@@ -6,7 +6,7 @@ import java.time.LocalDateTime;
 @Data
 public class ReportDTO {
     private Integer reportId;
-    private Integer adminUserId;
+    private Long adminUserId;
     private String reportType; // "service_demand", "user_activity", "platform_performance"
     private String data; // Có thể là JSON hoặc chuỗi dữ liệu
     private LocalDateTime generatedAt;
